Features/proxy generator (#193) (#194)
* Created proxy generator class

* Add additional connection proxy_gen

* Pass unittest, need more testing and cleaning

* Add module for ProxyGenerator

* Fix launch own process fail

* Fix warning tor, change ProxyGenerator init

* Update connection functions self-doc

* Update readme for proxy generator methods and usage

* Fix luminati naming and timeout issue

* Fix proxy generation in selenium webdriver and readme typo

* Add #162 request header change

* Add #187 author id addition change

* Update readme for author by id search and update test module

* Fix author id in bibtex

* Change release version to 0.5
