+++ scholarly/__init__.py
@@ -1,3 +1,3 @@
 from ._scholarly import _Scholarly
-
+from ._proxy_generator import ProxyGenerator
 scholarly = _Scholarly()
